patch fix (#7291)
* fix: digital options

* fix: breadcrumbs
