+++ src/features/components/templates/navigation/tab-nav-without-border/index.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState, useRef } from 'react'
-import { Breadcrumbs } from '@deriv-com/components'
 import { SwiperSlide, Swiper } from 'swiper/react'
 import SwiperCore, { Navigation } from 'swiper'
 import NavigationTabMenu from '../../tabs/navigation-menu'
@@ -47,19 +46,6 @@ const NavigationTabWithoutBorder = ({ tab_data }: NavigationTabWithoutBorderType
 
     return (
         <Container.Fluid pt="25x">
-            <Breadcrumbs
-                className="py-general-md mt-600 ml-400"
-                links={[
-                    {
-                        content: 'Home',
-                        href: '/',
-                    },
-                    {
-                        content: 'Options',
-                        href: '/trade-types/options',
-                    },
-                ]}
-            />
             <Flex.Box
                 padding_block="10x"
                 gap="6x"
